+++ R/aux_functions.R
@@ -1,7 +1,7 @@
 #Auxiliary functions; some from WeightIt
 
 #Function to process inputs and throw warnings or errors if inputs are incompatible with methods
-check.inputs <- function(mcall, method, distance, exact, mahvars, antiexact, caliper, discard, reestimate, s.weights, replace, ratio, m.order, estimand) {
+check.inputs <- function(mcall, method, distance, exact, mahvars, antiexact, caliper, discard, reestimate, s.weights, replace, ratio, min.controls = NULL, max.controls = NULL, m.order, estimand, ...) {
 
   null.method <- is.null(method)
   if (null.method) {
@@ -14,21 +14,21 @@ check.inputs <- function(mcall, method, distance, exact, mahvars, antiexact, cal
   ignored.inputs <- character(0)
   error.inputs <- character(0)
   if (null.method) {
-    for (i in c("exact", "mahvars", "antiexact", "caliper", "std.caliper", "replace", "ratio", "m.order")) {
+    for (i in c("exact", "mahvars", "antiexact", "caliper", "std.caliper", "replace", "ratio", "min.controls", "max.controls", "m.order")) {
       if (i %in% names(mcall) && !is.null(get0(i, inherits = FALSE))) {
         ignored.inputs <- c(ignored.inputs, i)
       }
     }
   }
   else if (method == "exact") {
-    for (i in c("distance", "exact", "mahvars", "antiexact", "caliper", "std.caliper", "discard", "reestimate", "replace", "ratio", "m.order")) {
+    for (i in c("distance", "exact", "mahvars", "antiexact", "caliper", "std.caliper", "discard", "reestimate", "replace", "ratio", "min.controls", "max.controls", "m.order")) {
       if (i %in% names(mcall) && !is.null(get0(i, inherits = FALSE))) {
         ignored.inputs <- c(ignored.inputs, i)
       }
     }
   }
   else if (method == "cem") {
-    for (i in c("distance", "exact", "mahvars", "antiexact", "caliper", "std.caliper", "discard", "reestimate", "replace", "ratio", "m.order")) {
+    for (i in c("distance", "exact", "mahvars", "antiexact", "caliper", "std.caliper", "discard", "reestimate", "replace", "ratio", "min.controls", "max.controls", "m.order")) {
       if (i %in% names(mcall) && !is.null(get0(i, inherits = FALSE))) {
         ignored.inputs <- c(ignored.inputs, i)
       }
@@ -82,9 +82,14 @@ check.inputs <- function(mcall, method, distance, exact, mahvars, antiexact, cal
         }
       }
     }
+    for (i in c("min.controls", "max.controls")) {
+      if (i %in% names(mcall) && !is.null(get0(i, inherits = FALSE))) {
+        ignored.inputs <- c(ignored.inputs, i)
+      }
+    }
   }
   else if (method == "cardinality") {
-    for (i in c("distance", "mahvars", "antiexact", "caliper", "std.caliper", "reestimate", "replace", "m.order")) {
+    for (i in c("distance", "mahvars", "antiexact", "caliper", "std.caliper", "reestimate", "replace", "min.controls", "m.order")) {
       if (i %in% names(mcall) && !is.null(get0(i, inherits = FALSE))) {
         ignored.inputs <- c(ignored.inputs, i)
       }
@@ -95,7 +100,7 @@ check.inputs <- function(mcall, method, distance, exact, mahvars, antiexact, cal
       stop("distance = \"mahalanobis\" is not compatible with subclassification.", call. = FALSE)
     }
 
-    for (i in c("exact", "mahvars", "antiexact", "caliper", "std.caliper", "replace", "ratio", "m.order")) {
+    for (i in c("exact", "mahvars", "antiexact", "caliper", "std.caliper", "replace", "ratio", "min.controls", "max.controls", "m.order")) {
       if (i %in% names(mcall) && !is.null(get0(i, inherits = FALSE))) {
         ignored.inputs <- c(ignored.inputs, i)
       }
@@ -119,7 +124,7 @@ check.inputs <- function(mcall, method, distance, exact, mahvars, antiexact, cal
 process.distance <- function(distance, method, treat) {
   if (is.null(distance) && !is.null(method)) stop(paste0("'distance' cannot be NULL with method = \"", method, "\"."), call. = FALSE)
   else if (is.character(distance) && length(distance) == 1) {
-    allowable.distances <- c("glm", "cbps", "gam", "mahalanobis", "nnet", "rpart", "bart", "randomforest")
+    allowable.distances <- c("glm", "cbps", "gam", "mahalanobis", "nnet", "rpart", "bart", "randomforest", "glmnet")
 
     if (tolower(distance) %in% c("cauchit", "cloglog", "linear.cloglog", "linear.log", "linear.logit", "linear.probit",
                         "linear.cauchit", "log", "probit")) {
@@ -149,7 +154,7 @@ process.distance <- function(distance, method, treat) {
   else if (!is.numeric(distance) || (!is.null(dim(distance)) && length(dim(distance)) != 2)) {
     stop("'distance' must be a string with the name of the distance measure to be used or a numeric vector or matrix containing distance measures.", call. = FALSE)
   }
-  else if (is.matrix(distance) && (is.null(method) || method %in% c("genetic", "subclass", "cem", "exact"))) {
+  else if (is.matrix(distance) && (is.null(method) || !method %in% c("nearest", "optimal", "full"))) {
     if (is.null(method)) method <- "NULL" else method <- paste0('"', method, '"')
     stop(paste0("'distance' cannot be supplied as a matrix with method = ", method, "."), call. = FALSE)
   }
@@ -180,23 +185,33 @@ process.distance <- function(distance, method, treat) {
 }
 
 #Function to check ratio is acceptable
-process.ratio <- function(ratio, min.controls = NULL, max.controls = NULL, na.ok = FALSE) {
+.process.ratio <- function(ratio = NULL, min.controls = NULL, max.controls = NULL, na.ok = FALSE) {
   if (length(ratio) == 0) ratio <- 1
-  if (!na.ok && anyNA(ratio)) stop("'ratio' cannot be NA.", call. = FALSE)
-  if (!anyNA(ratio) && (!is.atomic(ratio) || !is.numeric(ratio) || length(ratio) > 1 || ratio < 1)) {
+  ratio.na <- anyNA(ratio)
+  if (!na.ok && ratio.na) stop("'ratio' cannot be NA.", call. = FALSE)
+  if (!ratio.na && (!is.atomic(ratio) || !is.numeric(ratio) || length(ratio) > 1 || ratio < 1)) {
     stop("'ratio' must be a single positive number.", call. = FALSE)
   }
   if (is.null(max.controls)) {
     ratio <- round(ratio)
     return(c(ratio = ratio))
   }
+  else if (anyNA(max.controls) || !is.atomic(max.controls) || !is.numeric(max.controls) || length(max.controls) > 1) {
+    stop("'max.controls' must be a single positive number.", call. = FALSE)
+  }
+  else if (ratio.na) {
+
+  }
   else {
-    if (ratio <= 1) stop("'ratio' must be greater than 1 for variable ratio matching.", call. = FALSE)
+    if (!ratio.na && ratio <= 1) stop("'ratio' must be greater than 1 for variable ratio matching.", call. = FALSE)
 
     max.controls <- ceiling(max.controls)
     if (max.controls <= ratio) stop("'max.controls' must be greater than 'ratio' for variable ratio matching.", call. = FALSE)
 
     if (is.null(min.controls)) min.controls <- 1
+    else if (anyNA(max.controls) || !is.atomic(max.controls) || !is.numeric(max.controls) || length(max.controls) > 1) {
+        stop("'max.controls' must be a single positive number.", call. = FALSE)
+    }
     else min.controls <- floor(min.controls)
 
     if (min.controls < 1) stop("'min.controls' cannot be less than 1 for variable ratio matching.", call. = FALSE)
@@ -204,6 +219,72 @@ process.ratio <- function(ratio, min.controls = NULL, max.controls = NULL, na.ok
 
     return(c(ratio = ratio, min.controls = min.controls, max.controls = max.controls))
   }
+}
+process.ratio <- function(ratio = NULL, method, min.controls = NULL, max.controls = NULL, ...) {
+  #Should be run after process.inputs() and ignored inputs set to NULL
+  ratio.null <- length(ratio) == 0
+  ratio.na <- !ratio.null && anyNA(ratio)
+
+  if (method %in% c("nearest", "optimal")) {
+    if (ratio.null) ratio <- 1
+    else if (ratio.na) stop("'ratio' cannot be NA.", call. = FALSE)
+    else if (!is.atomic(ratio) || !is.numeric(ratio) || length(ratio) > 1 || ratio < 1) {
+      stop("'ratio' must be a single positive number greater than or equal to 1.", call. = FALSE)
+    }
+
+    if (is.null(max.controls)) {
+      ratio <- round(ratio)
+    }
+    else if (anyNA(max.controls) || !is.atomic(max.controls) || !is.numeric(max.controls) || length(max.controls) > 1) {
+      stop("'max.controls' must be a single positive number.", call. = FALSE)
+    }
+    else {
+      if (ratio <= 1) stop("'ratio' must be greater than 1 for variable ratio matching.", call. = FALSE)
+
+      max.controls <- ceiling(max.controls)
+      if (max.controls <= ratio) stop("'max.controls' must be greater than 'ratio' for variable ratio matching.", call. = FALSE)
+
+      if (is.null(min.controls)) min.controls <- 1
+      else if (anyNA(max.controls) || !is.atomic(max.controls) || !is.numeric(max.controls) || length(max.controls) > 1) {
+        stop("'max.controls' must be a single positive number.", call. = FALSE)
+      }
+      else min.controls <- floor(min.controls)
+
+      if (min.controls < 1) stop("'min.controls' cannot be less than 1 for variable ratio matching.", call. = FALSE)
+      else if (min.controls >= ratio) stop("'min.controls' must be less than 'ratio' for variable ratio matching.", call. = FALSE)
+    }
+  }
+  else if (method == "full") {
+    if (!ratio.null) stop("'ratio' cannot be specified with full matching.", call. = FALSE) #should never be called because of process.inputs()
+    if (!is.null(max.controls) && (anyNA(max.controls) || !is.atomic(max.controls) || !is.numeric(max.controls) || length(max.controls) > 1)) {
+      stop("'max.controls' must be a single positive number.", call. = FALSE)
+    }
+
+    if (!is.null(min.controls) && (anyNA(min.controls) || !is.atomic(min.controls) || !is.numeric(min.controls) || length(min.controls) > 1)) {
+      stop("'min.controls' must be a single positive number.", call. = FALSE)
+    }
+  }
+  else if (method == "genetic") {
+    if (ratio.null) ratio <- 1
+    else if (ratio.na) stop("'ratio' cannot be NA.", call. = FALSE)
+    else if (!is.atomic(ratio) || !is.numeric(ratio) || length(ratio) > 1 || ratio < 1) {
+      stop("'ratio' must be a single positive number greater than or equal to 1.", call. = FALSE)
+    }
+    ratio <- round(ratio)
+
+    min.controls <- max.controls <- NULL
+  }
+  else if (method == "cardinality") {
+    if (ratio.null) ratio <- 1
+    else if (!ratio.na && (!is.atomic(ratio) || !is.numeric(ratio) || length(ratio) > 1 || ratio < 0)) {
+      stop("'ratio' must be a single positive number or NA.", call. = FALSE)
+    }
+
+    min.controls <- max.controls <- NULL
+  }
+
+  return(list(ratio = ratio, min.controls = min.controls, max.controls = max.controls))
+
 }
 
 #Function to check if caliper is okay and process it
@@ -406,14 +487,14 @@ check.package <- function(package.name, alternative = FALSE) {
 }
 
 #Create info component of matchit object
-create_info <- function(method, fn1, link, discard, replace, ratio, mcall, mahalanobis, subclass, antiexact, distance_is_matrix) {
+create_info <- function(method, fn1, link, discard, replace, ratio, max.controls, mcall, mahalanobis, subclass, antiexact, distance_is_matrix) {
   info <- list(method = method,
                distance = if (is.null(fn1)) NULL else sub("distance2", "", fn1, fixed = TRUE),
                link = if (is.null(link)) NULL else link,
                discard = discard,
                replace = if (!is.null(method) && method %in% c("nearest", "genetic")) replace else NULL,
                ratio = if (!is.null(method) && method %in% c("nearest", "optimal", "genetic")) ratio else NULL,
-               max.controls = if (!is.null(method) && method %in% c("nearest", "optimal")) mcall[["max.controls"]] else NULL,
+               max.controls = if (!is.null(method) && method %in% c("nearest", "optimal")) max.controls else NULL,
                # mahalanobis = is.full.mahalanobis || !is.null(mahvars),
                mahalanobis = mahalanobis,
                # subclass = if (!is.null(method) && method == "subclass") length(unique(match.out$subclass[!is.na(match.out$subclass)])) else NULL,
@@ -794,7 +875,7 @@ mm2subclass <- function(mm, treat) {
 }
 
 #(Weighted) variance that uses special formula for binary variables
-wvar <- function(x, bin.var = NULL, w = NULL, na.rm = TRUE) {
+wvar <- function(x, bin.var = NULL, w = NULL) {
   if (is.null(w)) w <- rep(1, length(x))
   if (is.null(bin.var)) bin.var <- all(x == 0 | x == 1)
 
@@ -807,6 +888,27 @@ wvar <- function(x, bin.var = NULL, w = NULL, na.rm = TRUE) {
   else {
     #Reliability weights variance; same as cov.wt()
     sum(w * (x - mx)^2)/(1 - sum(w^2))
+  }
+}
+
+#Weighted mean faster than weighted.mean()
+wm <- function(x, w = NULL, na.rm = TRUE) {
+  if (is.null(w)) {
+    if (anyNA(x)) {
+      if (!na.rm) return(NA_real_)
+      nas <- which(is.na(x))
+      x <- x[-nas]
+    }
+    return(sum(x)/length(x))
+  }
+  else {
+    if (anyNA(x) || anyNA(w)) {
+      if (!na.rm) return(NA_real_)
+      nas <- which(is.na(x) | is.na(w))
+      x <- x[-nas]
+      w <- w[-nas]
+    }
+    return(sum(x*w)/sum(w))
   }
 }